# fix: Use .EqualFold() to parse urn prefixed UUIDs